requirements.txt
# Python packages for a local Python 3.8 virtualenv.
biopython==1.83
ete3==3.1.2
macsyfinder==2.0rc6
mdmparis-defense-finder==1.0.9
